requirements.txt
# API Server (Backend)
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Frontend Server
Flask==3.0.2
Jinja2==3.1.3
itsdangerous==2.1.2
Werkzeug==3.0.1
click==8.1.7

# Security / Utilities
certifi==2024.2.2